[Chore] 캘린더 셀 관련 수정 및 리팩토링
